backend/requirements.txt
flask
flask-sqlalchemy
python-dotenv
resend
psycopg2-binary
flask-cors
gunicorn